ändrat på plats vid val av inputs (1,2,3,4,5 för att gå till en metod) och la in dom i "MainMenu" class.. Program.cs är ganska tom! skapade även en class och metod "GoBackToMenu" som tar user tillbaka till menyn
