+++ Bank.Library/DatabaseHandler/DatabaseRepo.cs
@@ -15,6 +15,16 @@ namespace Bank.Library.DatabaseHandler
         static HashSet<Customer> _customerList = new HashSet<Customer>();
         static HashSet<Account> _accountList = new HashSet<Account>();
 
+        public static void GoBackToMenu()
+        {
+            Console.WriteLine("Press 9 to go back to menu");
+            var menu = Console.ReadLine();
+            if (menu == "9")
+            {
+                MainMenu.Menu();
+            }
+        }
+
         public static void ShowStats()
         {
             Console.WriteLine("Customers: " + _customerList.Count());
@@ -23,29 +33,32 @@ namespace Bank.Library.DatabaseHandler
         }
 
         public static void SearchCustomer()
-
         {
             Console.Clear();
             Console.WriteLine("Search for customer\n" + "You can use name or city.");
             var input = Console.ReadLine();
             bool noCustomersFound = true;
          
-                foreach (var item in Customers)
+            foreach (var item in Customers)
+            {
+                if (item.Name.ToUpper().Contains(input.ToUpper()) || item.City.ToUpper().Contains(input.ToUpper()))
                 {
-                    if (item.Name.ToUpper().Contains(input.ToUpper()) || item.City.ToUpper().Contains(input.ToUpper()))
-                    {
-                        Console.WriteLine("ID: " + item.Id + "\nName: " + item.Name + "\nCity " + item.City);
-                        noCustomersFound = false;
-                    }
+                    Console.WriteLine("ID: " + item.Id + "\nName: " + item.Name + "\nCity " + item.City);
+                    noCustomersFound = false;
                 }
-
-                if (noCustomersFound == true)
-                {
-                    Console.WriteLine("Can't find any customer with your input. Press enter to try again");
-
-                }
-                Console.ReadLine();
             }
+
+            if (noCustomersFound == true)
+            {
+                Console.WriteLine("Can't find any customer with your input. Press enter to try again");
+            }
+            Console.WriteLine("Press 9 to go back to menu");
+            var menu = Console.ReadLine();
+            if (menu == "9")
+            {
+                MainMenu.Menu();
+            }
+        }
 
         public static void ShowCustomerReport()
         {
